FS: remove ssize_t

+++ src/lib/fs/memoryfile.c
@@ -4,7 +4,7 @@
 #include "../fs.h"
 #include "../slice.h"
 
-static ssize_t File_memory_write(void *self, Slice(uint8_t) in);
+static size_t File_memory_write(void *self, Slice(uint8_t) in);
 
 static File_class File_memory = {
         .write = File_memory_write,
@@ -15,11 +15,11 @@ File *MemoryFile_new(Buffer *buf)
     return File_new(File_memory, buf, NULL, buf->_allocator);
 }
 
-static ssize_t File_memory_write(void *_self, Slice(uint8_t) in)
+static size_t File_memory_write(void *_self, Slice(uint8_t) in)
 {
     Buffer *self = _self;
     size_t n = Slice_size(&in);
     const uint8_t *d = _Slice_data(&in);
     _Vector_push(sizeof(uint8_t), self, n, d, n);
-    return (ssize_t) n;
+    return n;
 }
